+++ test/allocator/allocator-storage.ts
@@ -1,5 +1,5 @@
 import {DevProtocolInstance} from '../test-lib/instance'
-import {validateAddressErrorMessage} from '../test-lib/utils'
+import {validateAddressErrorMessage} from '../test-lib/utils/error'
 
 contract(
 	'AllocatorStorageTest',
