+++ test/common/storage/using-storage.ts
@@ -1,5 +1,5 @@
 import {UsingStorageTestInstance} from '../../../types/truffle-contracts'
-import {validateErrorMessage} from '../../test-lib/utils'
+import {validateErrorMessage} from '../../test-lib/utils/error'
 
 contract('UsingStorageTest', ([deployer]) => {
 	const UsingStorageTestContract = artifacts.require('UsingStorageTest')
