+++ test/dev/dev.ts
@@ -1,6 +1,6 @@
 import {DevProtocolInstance} from '../test-lib/instance'
 import {DevInstance} from '../../types/truffle-contracts'
-import {validateErrorMessage} from '../test-lib/utils'
+import {validateErrorMessage} from '../test-lib/utils/error'
 
 contract(
 	'Dev',
