+++ test/test-lib/const.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_ADDRESS = '0x0000000000000000000000000000000000000000'
+export const WEB3_URI = 'ws://localhost:7545'
