+++ test/test-lib/utils/error.ts
@@ -1,0 +1,13 @@
+export function validateErrorMessage(
+	result: any,
+	errorMessage: string,
+	reason = true
+): void {
+	const message = reason ? `Reason given: ${errorMessage}` : errorMessage
+	expect(result).to.be.an.instanceOf(Error)
+	expect((result as Error).message).to.include(message)
+}
+
+export function validateAddressErrorMessage(result: any, reason = true): void {
+	validateErrorMessage(result, 'this is illegal address', reason)
+}
